imp: refactor validation error formatting
